# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Apex Health ecommerce API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog
  - name: checkout
    description: Checkout and payments
  - name: orders
    description: Order management
  - name: customers
    description: Customer accounts
  - name: admin
    description: Admin operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List active products with prices
      parameters:
        - in: query
          name: category
          schema:
            type: string
          required: false
        - in: query
          name: limit
          schema:
            type: integer
          required: false
        - in: query
          name: offset
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"

  /products/featured:
    get:
      operationId: listFeaturedProducts
      tags: [products]
      summary: List featured products for homepage
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product with prices
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /checkout:
    post:
      operationId: createCheckout
      tags: [checkout]
      summary: Create a Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID (for confirmation page)
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /customers:
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Register a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /customers/{email}/orders:
    get:
      operationId: getCustomerOrders
      tags: [customers]
      summary: Get orders for a customer by email
      parameters:
        - in: path
          name: email
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Customer orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"

  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      summary: List all products (including inactive)
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          required: false
        - in: query
          name: offset
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: All products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: adminCreateProduct
      tags: [admin]
      summary: Create a product in Stripe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/products/{id}:
    patch:
      operationId: adminUpdateProduct
      tags: [admin]
      summary: Update a product
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeactivateProduct
      tags: [admin]
      summary: Deactivate a product
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Product deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/prices:
    get:
      operationId: adminListPrices
      tags: [admin]
      summary: List all prices
      parameters:
        - in: query
          name: productId
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Prices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceListResponse"
    post:
      operationId: adminCreatePrice
      tags: [admin]
      summary: Create a price for a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceInput"
      responses:
        "201":
          description: Price created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Price"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/prices/{id}:
    delete:
      operationId: adminDeactivatePrice
      tags: [admin]
      summary: Deactivate a price
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Price deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Price"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: List all orders
      parameters:
        - in: query
          name: status
          schema:
            type: string
          required: false
        - in: query
          name: limit
          schema:
            type: integer
          required: false
        - in: query
          name: offset
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"

  /admin/orders/{id}:
    get:
      operationId: adminGetOrder
      tags: [admin]
      summary: Get order detail
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: adminUpdateOrder
      tags: [admin]
      summary: Update order status
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          required: false
        - in: query
          name: offset
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"

  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Get user detail
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    Price:
      type: object
      required: [id, unitAmount, currency, active, productId]
      properties:
        id:
          type: string
        unitAmount:
          type: integer
        currency:
          type: string
        active:
          type: boolean
        productId:
          type: string

    Product:
      type: object
      required: [id, name, active]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        active:
          type: boolean
        category:
          type: ["string", "null"]
        imageKey:
          type: ["string", "null"]
        featured:
          type: boolean
        prices:
          type: array
          items:
            $ref: "#/components/schemas/Price"

    ProductListResponse:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer

    ProductInput:
      type: object
      required: [name, unitAmount]
      properties:
        name:
          type: string
        description:
          type: string
        unitAmount:
          type: integer
        currency:
          type: string
        category:
          type: string
        imageKey:
          type: string
        featured:
          type: boolean

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
        category:
          type: string
        imageKey:
          type: string
        featured:
          type: boolean

    PriceInput:
      type: object
      required: [productId, unitAmount]
      properties:
        productId:
          type: string
        unitAmount:
          type: integer
        currency:
          type: string

    PriceListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Price"

    CheckoutInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        customerEmail:
          type: string
        successUrl:
          type: string
        cancelUrl:
          type: string

    CartItem:
      type: object
      required: [priceId, quantity]
      properties:
        priceId:
          type: string
        quantity:
          type: integer

    CheckoutResponse:
      type: object
      required: [url, sessionId]
      properties:
        url:
          type: string
        sessionId:
          type: string

    OrderItem:
      type: object
      required: [id, productId, productName, unitAmount, quantity, currency]
      properties:
        id:
          type: integer
        productId:
          type: string
        productName:
          type: string
        priceId:
          type: ["string", "null"]
        unitAmount:
          type: integer
        quantity:
          type: integer
        currency:
          type: string

    Order:
      type: object
      required: [id, customerEmail, status, totalAmount, currency, createdAt]
      properties:
        id:
          type: integer
        stripeSessionId:
          type: ["string", "null"]
        stripePaymentIntentId:
          type: ["string", "null"]
        customerEmail:
          type: string
        customerName:
          type: ["string", "null"]
        status:
          type: string
        totalAmount:
          type: integer
        currency:
          type: string
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    OrderUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string

    OrderListResponse:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer

    CustomerInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
        name:
          type: string

    Customer:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        stripeCustomerId:
          type: ["string", "null"]
        createdAt:
          type: string

    User:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        stripeCustomerId:
          type: ["string", "null"]
        totalOrders:
          type: integer
        totalSpent:
          type: integer
        createdAt:
          type: string

    UserListResponse:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer

    AdminStats:
      type: object
      required: [totalOrders, totalRevenue, totalCustomers, recentOrders]
      properties:
        totalOrders:
          type: integer
        totalRevenue:
          type: integer
        totalCustomers:
          type: integer
        newOrdersThisWeek:
          type: integer
        revenueThisWeek:
          type: integer
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/Order"