# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Apex Health ecommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: number;
  /** @nullable */
  stripeSessionId?: string | null;
  /** @nullable */
  stripePaymentIntentId?: string | null;
  customerEmail: string;
  /** @nullable */
  customerName?: string | null;
  status: string;
  totalAmount: number;
  currency: string;
  createdAt: string;
  items?: OrderItem[];
}