# lib/api-zod/src/generated/types/priceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Apex Health ecommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface PriceInput {
  productId: string;
  unitAmount: number;
  currency?: string;
}